Rename finality icon
