+++ packages/frontend2/src/icons/pages/finality.tsx
